requirements.txt
decord==0.6.0
h5py==3.14.0
hydra-core==1.3.2
iopath==0.1.10
matplotlib==3.3.4
maxflow==0.0.1
numpy==1.19.5
omegaconf==2.3.0
open3d==0.19.0
Pillow==11.3.0
pycocotools==2.0.10
pymeshfix==0.17.1
pytorch3d==0.7.4
scikit_learn==1.7.1
scipy==1.16.1
stl==0.0.3
torch==2.7.1
torchvision==0.22.1
tqdm==4.64.1
